1、多页跳转到聊天详情界面，处理聊天详情界面的返回逻辑，让其始终跳转到聊天列表 2、常见问题与个人资料接口调用调试，常见问题布局调整，展开列表 3、定位数据处理，可能出现第一次定位获取不到城市，采用连续定位，直到获取到城市信息进行停止定位，进行打卡状态的请求

+++ app/src/main/java/com/example/smartagriculture/viewmodel/WarnMessageViewModel.kt
@@ -33,7 +33,7 @@ class WarnMessageViewModel(application: Application,
     fun getWarnList( beanDataList: BeanDataList,parkId:String): Unit {
         var commitParam= CommitParam()
         commitParam.parkId = parkId
-        commitParam.warntype = beanDataList.id
+        commitParam.warntype = beanDataList.id.toInt()
         commitParam.page = "1"
         commitParam.pageSize = "10"
         var  map=HashMap<String,String>()
